backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.4.0
pymongo==4.6.3
pydantic==2.9.2
pydantic-settings==2.9.1
python-dotenv==1.0.1
python-jose==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0
email-validator==2.3.0
kubernetes==28.1.0
PyYAML==6.0.1
openstacksdk==3.1.0
aiofiles>=25.1.0
python-multipart>=0.0.20
cryptography==43.0.1

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
requests>=2.31.0
httpx>=0.27.0
